gh-115: Incapsulate lexical_grammar.Ecma262Parser
The whole PR revolves around this diff:

```diff
 #[derive(Parser)]
 #[grammar = "lexical_grammar.pest"]
-pub struct Ecma262Parser;
+struct Ecma262Parser;
```

The rest of the PR is accomodation to this.
