+++ tests/test_tokenizer.rs
@@ -1,75 +1,73 @@
 #[cfg(test)]
 mod tests {
     use claims::{assert_err, assert_matches, assert_ok_eq};
-    use embedded_ecmascript::{
+    use embedded_ecmascript::lexical_grammar::{
+        Addition,
+        AdditionAssignment,
+        And,
+        AndAssignment,
+        Assignment,
+        BitAnd,
+        BitAndAssignment,
+        BitNot,
+        BitOr,
+        BitOrAssignment,
+        BitXor,
+        BitXorAssignment,
+        ClosingBracket,
+        ClosingParenthesis,
+        Colon,
+        Comma,
+        CommonToken,
+        Decrement,
+        Division,
+        DivisionAssignment,
+        DivPunctuator,
+        Dot,
+        Ellipsis,
+        Exponentiation,
+        ExponentiationAssignment,
+        FunctionArrow,
         get_next_token,
         GoalSymbols,
-        lexical_grammar::{
-            Addition,
-            AdditionAssignment,
-            And,
-            AndAssignment,
-            Assignment,
-            BitAnd,
-            BitAndAssignment,
-            BitNot,
-            BitOr,
-            BitOrAssignment,
-            BitXor,
-            BitXorAssignment,
-            ClosingBracket,
-            ClosingParenthesis,
-            Colon,
-            Comma,
-            CommonToken,
-            Decrement,
-            Division,
-            DivisionAssignment,
-            DivPunctuator,
-            Dot,
-            Ellipsis,
-            Exponentiation,
-            ExponentiationAssignment,
-            FunctionArrow,
-            Increment,
-            LeftShift,
-            LeftShiftAssignment,
-            Less,
-            LessOrEqual,
-            LineTerminator,
-            LooseEquality,
-            LooseInequality,
-            Modulo,
-            ModuloAssignment,
-            More,
-            MoreOrEqual,
-            Multiplication,
-            MultiplicationAssignment,
-            Not,
-            NullishCoalescence,
-            NullishCoalescenceAssignment,
-            OpeningBrace,
-            OpeningBracket,
-            OpeningParenthesis,
-            OptionalChainingPunctuator,
-            Or,
-            OrAssignment,
-            OtherPunctuator,
-            Punctuator,
-            RightBracePunctuator,
-            RightShift,
-            RightShiftAssignment,
-            StrictEquality,
-            StrictInequality,
-            SubtractionAssignment,
-            QuestionMark,
-            Semicolon,
-            Subtraction,
-            UnsignedRightShift,
-            UnsignedRightShiftAssignment,
-            WhiteSpace,
-        },
+        Increment,
+        LeftShift,
+        LeftShiftAssignment,
+        Less,
+        LessOrEqual,
+        LineTerminator,
+        LooseEquality,
+        LooseInequality,
+        Modulo,
+        ModuloAssignment,
+        More,
+        MoreOrEqual,
+        Multiplication,
+        MultiplicationAssignment,
+        Not,
+        NullishCoalescence,
+        NullishCoalescenceAssignment,
+        OpeningBrace,
+        OpeningBracket,
+        OpeningParenthesis,
+        OptionalChainingPunctuator,
+        Or,
+        OrAssignment,
+        OtherPunctuator,
+        Punctuator,
+        RightBracePunctuator,
+        RightShift,
+        RightShiftAssignment,
+        StrictEquality,
+        StrictInequality,
+        SubtractionAssignment,
+        QuestionMark,
+        Semicolon,
+        Subtraction,
         Token,
+        UnsignedRightShift,
+        UnsignedRightShiftAssignment,
+        WhiteSpace,
     };
     use rstest::rstest;
